Refactor GeoTIFF Phase 5c-read: extract _layout.py from _reader.py (#2247)
Move read-side layout / validation helpers out of ``_reader.py`` into a
new ``_layout`` module so the top-level reader keeps just orchestration
and the COG-HTTP transport layer.

Moved:
- ``MAX_PIXELS_DEFAULT``, ``PixelSafetyLimitError``,
  ``_check_dimensions``, ``_check_source_dimensions``
- ``_sparse_fill_value``, ``_has_sparse``
- ``_FULL_IMAGE_BUDGET_HEADER_SLACK``, ``_compute_full_image_byte_budget``
- ``_ifd_required_extent``

``_reader.py`` re-exports each name so existing
``from xrspatial.geotiff._reader import ...`` call sites and test
monkeypatches against ``_reader.MAX_PIXELS_DEFAULT`` keep working.
``_decode.py`` switches to importing the helpers from ``_layout``
directly and keeps the ``_reader.MAX_PIXELS_DEFAULT`` lookup so the
existing monkeypatch contract on the per-tile safety check still
fires.

Mechanical, behaviour-neutral. ``_reader.py`` drops from 1406 to 1258
lines; the remaining bulk is the COG-HTTP transport which the issue
scopes out of this PR.
